pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[project]
name = "app"
version = "0.0.1"

description = "FastApi Polygon Proxy"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastapi==0.100.*",
    "uvicorn==0.23.*",
    "sqlalchemy==2.*",
    "alembic==1.11.*",
    "psycopg2-binary",
    "python-multipart",
    "python_dotenv",
    "celery",
    "motorengine"
]
[project.optional-dependencies]
test = [
    "pytest"
]
lint = [
    "ruff",
    "black",
]
